move sections into a section/ folder
