+++ src/pages/index.js
@@ -2,9 +2,9 @@ import React from "react"
 
 import Page from "../components/page"
 import Section from "../components/section"
-import Intro from "../components/intro"
+import Intro from "../sections/intro"
 import Spacer from "../components/spacer"
-import Explanation from "../components/explanation"
+import Explanation from "../sections/explanation"
 
 const IndexPage = () => (
   <Page title="JSCraftCamp">
