+++ src/pages/sponsoring.js
@@ -1,7 +1,7 @@
 import React from "react"
 
 import Page from "../components/page"
-import ExplainSponsoring from "../components/explain-sponsoring"
+import ExplainSponsoring from "../sections/explain-sponsoring"
 
 export default function Sponsoring() {
   return (
