+++ src/pages/values.js
@@ -1,6 +1,6 @@
 import React from "react"
 import Page from "../components/page"
-import CodeOfConduct from "../components/code-of-conduct"
+import CodeOfConduct from "../sections/code-of-conduct"
 
 const Values = () => (
   <Page title="Values">
