+++ src/sections/explanation/explanation.module.scss
