+++ src/sections/explanation/index.js
